+++ hal/src/thumbv7em/clock/v2/gclkio.rs
@@ -60,7 +60,7 @@ use core::marker::PhantomData;
 use seq_macro::seq;
 use typenum::U0;
 
-use crate::clock::v2::{types::Enabled, Source, SourceMarker};
+use crate::clock::v2::{Enabled, Source, SourceMarker};
 use crate::gpio::v2::{self as gpio, AlternateM, AnyPin, Pin, PinId};
 use crate::time::Hertz;
 use crate::typelevel::{Counter, Decrement, Increment, Sealed};
