+++ hal/src/thumbv7em/clock/v2/osculp32k.rs
@@ -20,7 +20,7 @@ use typenum::U0;
 use crate::pac::osc32kctrl::rtcctrl::RTCSEL_A;
 use crate::pac::osc32kctrl::OSCULP32K;
 
-use crate::clock::v2::{rtc::RtcSourceMarker, types::Enabled, Source, SourceMarker};
+use crate::clock::v2::{rtc::RtcSourceMarker, Enabled, Source, SourceMarker};
 use crate::time::{Hertz, U32Ext};
 use crate::typelevel::{Counter, Sealed};
