+++ hal/src/thumbv7em/clock/v2/types.rs
@@ -254,45 +254,3 @@
 //! different tokens do not overlap. Tokens cannot be created by a user; they
 //! are provided during initialization and do not expose any public API. Memory
 //! accesses are read/write-synchronized.
-
-use typenum::Unsigned;
-
-use crate::typelevel::{Counter, Decrement, Increment, PrivateDecrement, PrivateIncrement, Sealed};
-
-/// Wrapper type representing enabled multiuser clocking abstraction
-///
-/// Specialized implementations for this type usually allow for a specific
-/// behaviour dependent on a user count. For instance, clocking components are
-/// disableable only when no one is using them.
-pub struct Enabled<T, N: Counter>(pub(crate) T, N);
-
-impl<T, N: Counter> Sealed for Enabled<T, N> {}
-
-impl<T, N: Unsigned + Counter> Enabled<T, N> {
-    #[inline]
-    pub(crate) fn new(t: T) -> Self {
-        Enabled(t, N::default())
-    }
-}
-
-impl<T, N: PrivateIncrement> PrivateIncrement for Enabled<T, N> {
-    type Inc = Enabled<T, N::Inc>;
-
-    #[inline]
-    fn inc(self) -> Self::Inc {
-        Enabled(self.0, self.1.inc())
-    }
-}
-
-impl<T, N: PrivateDecrement> PrivateDecrement for Enabled<T, N> {
-    type Dec = Enabled<T, N::Dec>;
-
-    #[inline]
-    fn dec(self) -> Self::Dec {
-        Enabled(self.0, self.1.dec())
-    }
-}
-
-impl<T, N: PrivateIncrement> Increment for Enabled<T, N> {}
-impl<T, N: PrivateDecrement> Decrement for Enabled<T, N> {}
-impl<T, N: Counter> Counter for Enabled<T, N> {}
